+++ service/brookesia_service_wifi/include/brookesia/service_wifi.hpp
@@ -6,6 +6,6 @@
 #pragma once
 
 #include "service_wifi/macro_configs.h"
-#include "service_wifi/hal.hpp"
+#include "service_wifi/hal/hal.hpp"
 #include "service_wifi/state_machine.hpp"
 #include "service_wifi/service_wifi.hpp"